requirements.txt
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.14.0
pandas==2.2.3
requests==2.32.3
pyarrow==17.0.0
google-cloud==0.34.0
numpy==1.26.2